✨ feat: Stepper 컴포넌트를 EventStepper로 변경 및 불필요한 props 제거
